third_party/WebKit/LayoutTests/fast/viewport/viewport-79-expected.txt

CONSOLE MESSAGE: line 3: The key "a" is not recognized and ignored.
CONSOLE MESSAGE: line 3: Error parsing a meta element's content: ';' is not a valid key-value pair separator. Please use ',' instead.
ALERT: viewport size 320x352 scale 1 with limits [1, 5] and userScalable true
